Introduced accessibility for models
A view model has now `canAccess` which should determine, whether the
operator can access the model. Currently, only ViewMeeting is affected:
We need to hide those meetings with too less information. This happens
due to the fact, that each meeting is always given to a user (even
anonymous) to provide basic infos, e.g. for login.

These meetings must not be shown in lists and the normal UI. The base
repo only outputs accessible models by default. There are some new
"unsafe" accessors to query unaccessible models. This will be needed by
the per-meeting-login mask.

Additionally moved some misplaced files and worked on the meeting edit.
It is not finished, but wrong parts are removed so far. A good
update-method for the server is missing.