Poll Slide
Unified the usage of `PollData`: The poll services and
display-components only work with PollData, not ViewPoll. ViewPoll as
well as the data for the projector interfaces with PollData, so PollData
is the only single interface to operate on.

Also cleaned up the assignment/motion-poll service; Since there is just
one poll model in OS4, the data is much more streamedlined, so much code
could be moved into the poll service.

The slide for analog polls works; Maybe something else broke, but this
PR is more of a cleanup-type. Building upon this PR, the bugs/missing
logic can be fixed/implemented in a more clean fashion.