Improves the "cleanup" command for the client
At the moment, the command "npm run cleanup" is not practical. The command executes "npm run prettify-write" first and then "npm run lint-write". However, after executing "npm run lint-write" some changes are not "beautiful" enough to go through the "prettify-check". Therefore, it is necessary to execute the command "npm run lint-write" before executing "npm run prettify-write".
The commands are executed in the previously mentioned order, because "npm run prettify-write" floods a console. To avoid this, I wrote the script "cleanup.ts".
This should also have the benefit, that we do not need the command "cleanup-win".