Rework of the whole client (#989)
* Rework of the whole client

- This will reduce the necessary bandwidth to use OpenSlides (Fixes #861)
- This will also reduce the whole image size of OpenSlides and introduce base methods to make it smaller (Fixes #860)
- Update the app to Angular@13 (Fixes #990)
- Also updates other dependencies, so there are no security issues
- Prevents using `getViewModelListBehaviorSubject` in a repo (Fixes #212)
- Updates the used image to Node:16.10 (Fixes #393)
- Looked into every file and removed unused files (Fixes #339)
- Will achieve the handling of accounts in the account-list as discussed (Fixes #542)
- Implements a search selector for accounts (Fixes #665)
- There is no need to figure out which date of a model is at which time necessary (Fixes #900)
- Use newer icons (Fixes #876)

There are also several other changes:
Every file (component, service and other things) is provided by a module. This has the effect, that the client does not load every used library on its initial loading. There are also some basic files which will be used for other changes/issues (for example #124). Most of the pages in the application are splitted into two/three parts: a wrapper to load necessary data for a specific page (especially "detail" pages), a "view" and an "edit" part. Separates dialogs into dedicated modules. Splits the structure into a more domain driven one: domain, gateways, infrastructure, site and ui. Gateways are the "gates" to the "outer" world of the client (other services/the backend, user devices...). The ui contains custom ui-elements. The site contains every page and is organized in three main parts: the login, the organization and the meetings. Gateways contain repositories, which are only used to communicate with the backend and holding models in their dedicated collection. There are also controllers, which are separated from the repositories to make "manipulations" for their dedicated collection or custom requests. The domain contains every domain specific file, which means the models/collections and some interfaces. There are also actions which provide additional functionality for sending requests to the backend.