Make text of dropZoneLabel wrappable
After applying paddings, the ending of the message was no longer be able to
fit on the small screens (like iPhone SE).

This commit sets 'display: flex' for the container around dropZoneLabel and
changes the line heigth of a dropZoneLabel to 24px (which is 150% of the
font size). These chages let the text be fully visible on the small screens.